get rid of tabulardataset and replace with set of utility functions
