+++ vulcanai/datasets/__init__.py
@@ -13,12 +13,11 @@ Submodules
 """
 
 from .fashion import FashionData
-from .tabulardataset import TabularDataset
 from .multidataset import MultiDataset
 
 __all__ = [
     'fashion',
-    'tabulardataset',
+    'tabular_data_utils.py',
     'utils',
     'FashionData',
     'TabularDataset',
