+++ vulcanai/datasets/utils.py
@@ -106,78 +106,4 @@ def rationed_split(df, train_ratio, test_ratio, validation_ratio):
     return indices
 
 
-def stitch_datasets(df_main=None, merge_on_columns=None,
-                    index_list=None, **dataset_dict):
-    """
-    Function to produce a single dataset from multiple.
 
-    Parameters:
-        df_main: dataframe
-            optional primary dataframe to merge onto
-        merge_on_columns : list of strings
-            key(s) that specifies which columns to use to uniquely stitch
-            dataset (default None)
-        index_list: list of strings
-            columns to establish as index for final stitched dataset
-            (default None)
-        dataset_dict : keyword parameter, value is dataframe
-            pandas dataframe assigned to keyword argument that produces a
-            dictionary variable.
-
-    Returns:
-        merged_df : dataframe
-            concatenated dataframe
-
-    :Example:
-
-    >>> df_test_one: pd.DataFrame({'A': ['A0', 'A1', 'A2', 'A3'],
-                                   'B': ['B0', 'B1', 'B2', 'B3'],
-                                   'C': ['C0', 'C1', 'C2', 'C3'],
-                                   'D': ['D0', 'D1', 'D2', 'D3']},
-                                   index=[0, 1, 2, 3]),
-
-     >>> df_test_two: pd.DataFrame({'A': ['A4', 'A5', 'A6', 'A7'],
-                                    'B': ['B4', 'B5', 'B6', 'B7'],
-                                    'C': ['C4', 'C5', 'C6', 'C7'],
-                                    'D': ['D4', 'D5', 'D6', 'D7']},
-                                    index=[4, 5, 6, 7])}
-    >>> df_stitched = stitch_datasets(merge_on_columns=['A'], index_list=['A'],
-     df1=df_test_one, df2=df_test_two)
-
-    """
-    # Get name of a dataframe to extract
-    if df_main is not None:
-        merged_df = copy.deepcopy(df_main)
-    else:
-        first_column = list(dataset_dict)[0]
-        merged_df = dataset_dict.pop(first_column)
-        merged_df = merged_df.apply(pd.to_numeric, errors='ignore')
-
-    for key in list(dataset_dict):
-        logger.info('Combining: {}'.format(key))
-        df_two = dataset_dict.pop(key)
-        merged_df = pd.concat([merged_df, df_two], sort=False)
-
-    if merge_on_columns is not None:
-        # Group by keys, forward fill and backward fill missing data
-        # then remove duplicate keys
-        merged_df = merged_df.apply(pd.to_numeric, errors='ignore')
-        df_group_on = merged_df.reset_index(drop=True).\
-            groupby(merge_on_columns).apply(lambda x: x.bfill().ffill())
-        logger.info("\tDropping duplicates")
-
-        # Drop rows where there are duplicates for the merged_on_columns.
-        # We first need to dropna based on merged since drop_duplicates
-        # ignores null/na values.
-        df_group_on = df_group_on.dropna(subset=merge_on_columns, how='all')
-        df_group_on = df_group_on.drop_duplicates(subset=merge_on_columns,
-                                                  keep='first', inplace=False)
-        merged_df = df_group_on
-
-    if index_list is not None:
-        merged_df = merged_df.set_index(index_list, inplace=False)
-
-    logger.info("\nMerge Total columns = {totalCols}, rows = {totalRows} "
-                .format(totalCols=len(list(merged_df)),
-                        totalRows=len(merged_df)))
-    return merged_df
